fix: panic when using pointer to map as env
Previously, passing a pointer to a map (e.g. `*map[string]any`)
to `expr.Env()` would cause a panic because reflection methods
were called on the pointer value instead of the underlying map.
With this change the value is first deref'd in the config before
any map operations are performed.

Consequently, the `vm.Run()` method needs to dereference the env
map. This ensures compatibility with the optimisations like
`OpLoadFast` where the compiler determines the environment simply
as `map[string]any`.

Regression test added.

Signed-off-by: Ville Vesilehto <ville@vesilehto.fi>